Refactor proof module layout
